build: Add PIC/PIE support (only works on linuxu target)
Use -fPIC and -pie as compiler and linker flags to enable PIC (position
independent code) and PIE (position independent executable). This
currently only works on linuxu target.

The KVM target requires the --omagic flag that conflicts with --pie.
--omagic is required by the .data.boot section. For KVM more than one
thing needs to be tackled:
* use PIC code in KVM platform assembly files (such as entry64.S)
* disable the requirement of --omagic; this will require tampering with
the .data.boot section in entry64.S
* create a source of randomization during runtime

Signed-off-by: Razvan Deaconescu <razvan.deaconescu@cs.pub.ro>