plat/common: Introduce page tables API
Here are the functions used to manipulate memory mappings, agnostic of
architecture and platform. Since the strucutre of the page tables is
hierarchical with 4 levels on both x86_64 and ARM64, the logic for page
table walking is the same.

Differences are in the abstracted ukarch_* functions, which were
introduced in the previous patches for the x86_64 architecture.

uk_pt_init() function sets up the memory layout, which keeps the
internal bookkeeping. This layout migh suffer changes later, when this
API will be integrated with the memory layout of the whole system. Right
now, the components are:
- a bitmap for physical addresses which are available to be mapped to
- a memory range for storing the actual page tables
- a bitmap for bookkeeping the page table area

Signed-off-by: Stefan Teodorescu <stefanl.teodorescu@gmail.com>