lib/ukboot: DON'T UPSTREAM Init pagetables at boot
This call "steals" half of the memory given to the allocator at boot
time, so it considers it as the whole physical memory available to map
at. I do this split in half, so I make sure that when a page is mapped
on a physical frame within this area.

The allocator shouldn't return a virtual address which is already mapped
to a physical frame returned by my mmap implementation or used
internally by my implementation.

Signed-off-by: Stefan Teodorescu <stefanl.teodorescu@gmail.com>