arch/*, build: Add support for clang compilation
In order to perform the compilation and linking with clang without warnings, the following flags must be removed:
-fno-tree-sra, -fms-extensions, -nostdinc, -fno-reorder-blocks, -mno-fp-ret-in-387. Functions such as have_clang,
have_gcc, clang_version_ge, error_if_clang_version_lt have been added to solve the problem.
With all of this we are able to rearrange the flags.
So, if there are flags that fail with clang, we can select them only for GCC compilation.
Also, we can introduce specific flags for clang.

Signed-off-by: Alice Suiu <alicesuiu17@gmail.com>