build: Add PGO support
Profile-Guided Optimization is a method used by compilers to produce
optimal code by using application runtime data. PGO involves two
phases one that implies a compilation and a run that will generate
the profiling files for the application and one that implies a
compilation in which the compiler uses the profiling data to make a
better estimate of the program’s control flow and to produce an optimized
executable file.

Add the configuration for PGO phases in MENUCONFIG and the correspondent
flag for LDFLAGS.

Signed-off-by: Alice Suiu <alicesuiu17@gmail.com>