plat/kvm: Change the default directory for the profiling files
Call the __gcov_exit() function in a conditional way (using ifdef guard) in
the function ukplat_terminate(). Thus, the profiling files (*.gcda) are generated
within the Unikraft filesystem. Use 9pfs to transfer the *.gcda files from Unikraft
to the host system.

Set GCOV_PREFIX and GCOV_PREFIX_STRIP. GCOV_PREFIX is a runtime environment variable
that determines the root directory where the gcov output files (*.gcda) are written.
GCOV_PREFIX_STRIP=X is also a runtime variable and has the effect of stripping X
elements from the path found in the object files.

Thus, the default path in which the *.gcda files were generated was changed with the
help of the two environment variables. Now the files are saved to the path given
exclusively by the GCOV_PREFIX environment variable, GCOV_PREFIX_STRIP being large
enough to skip the entire default path.

Signed-off-by: Alice Suiu <alicesuiu17@gmail.com>