plat/common: Introduce page tables API
Here are the functions used to manipulate memory mappings, agnostic of
architecture and platform. Since the strucutre of the page tables is
hierarchical with 4 levels on both x86_64 and ARM64, the logic for page
table walking is the same.

Differences are in the abstracted ukarch_* functions, which were
introduced in the previous patches for the x86_64 architecture.

uk_pt_init() function sets up the internal bookkeeping of the API and
receives as parameters the virtual address of where the page tables
should be placed in memory and the physicall address and length of the
memory that should be managed by the API. The API can be initialized
without any free physical memory to be managed and can be used to
manipulate the existing mappings.

The internal structures are:
- a bitmap for the available physical frames
- a bitmap for the internal page tables area
- a memory range where page tables are stored

Signed-off-by: Stefan Teodorescu <stefanl.teodorescu@gmail.com>