plat/common: Add map/unmap functions to page table API
uk_page_map() and uk_page_unmap() are the functions that map and unmap
pages (receiving a virtual address as the parameter). For now, these
only handle mapping and unmapping one page at a time. The caller of
these functions need not be aware of the physical addresses used by the
mappings, but the option to supply specific physical addresses is
available (for example, for memory mapped I/O, or page grants on Xen).

Signed-off-by: Stefan Teodorescu <stefanl.teodorescu@gmail.com>