include: Introduce virtual memory layout
This file contains the virtual memory regions that need to be
initialized at boot time when building the page table. These regions
include: the kernel, the area for the page tables, area for heap, stack,
for mappings using mmap etc.

These abstractions also aid in introducing ASLR and PIE, which will only
need to change the start address of specific regions.

Signed-off-by: Stefan Teodorescu <stefanl.teodorescu@gmail.com>