plat/common: Introduce functions for page table boot-time initialization
Introduce functions for boot-time initialization of page tables. These
functions include the creation of a new page table from scratch, mapping
the kernel, the heap and the stack.

There are two ways the page table API can be used:

1) In conjunction with the linear 1:1 mapping page table that is
currently used. This mode allows, to some extent, modifications of existing
mappings and creation of new ones. For example, on KVM, the static page
table only maps the first 1G of memory. If the VM is started with more
than 1G of memory, this extra memory can be managed using the API. If
there is no extra memory (above 1G), the API can be used to modify
existing mappings.

2) Completely build a new page table, from scratch, at boot-time. This
allows more flexibility and introduction of security mechanisms, such as
ASLR and PIE by creating all mappings for all areas.

At the moment, the first mode is available on both KVM and Xen
platforms, and is usable by calling the uk_pt_init() function. The
second mode, the dynamic initialization, is available on KVM, by calling
the uk_pt_build() function.

Signed-off-by: Stefan Teodorescu <stefanl.teodorescu@gmail.com>