plat/kvm: Enable page table API and dynamic initialization
If the dynamic initialization of page tables configuration is selected,
the memory layout at include/uk/arch/mem_layout.h is used. If POSIX mmap
functions are enabled, the heap memory is divided in half between the
heap and the memory available for mmap. This is a temporary solution,
and should be replaced by implementing a mechanism for dynamically
changing the heap size (e.g. brk()).

If the page table API configuration is selected, but not the dynamic
initialization, the current page table and current layout are kept, but
theere is the posibility to manage any memory above 1G and to modify
existing mappings, below 1G.

Signed-off-by: Stefan Teodorescu <stefanl.teodorescu@gmail.com>