plat/xen: Enable page table API
At the moment, on Xen, the dynamic initialization of page tables is not
supported. However, the virtual memory management API can still be used
to manage existing mappings and to use POSIX mmap.

Depending on whether POSIX mmap is selected or not, the heap is either
halved (to be split between heap and memory available for mmap use,
similar to KVM) or left as it is.

Signed-off-by: Stefan Teodorescu <stefanl.teodorescu@gmail.com>