plat/common: Introduce stack allocation functions
When using dynamically initialized virtual memory layout, multiple
stacks may be used when there are multiple threads (number of stacks ==
number of threads).

For the purpose of allocating a new stack, the function uk_stack_alloc()
can be used, which returns the lower address of the new stack.
For example: void *addr = uk_stack_alloc();. In this case, the allocated
stack is the segment [addr, addr + _STACK_SIZE].

When a stack is no longer needed (e.g. when a thread finishes execution),
uk_stack_free(addr) is called to free its stack.

Signed-off-by: Stefan Teodorescu <stefanl.teodorescu@gmail.com>