plat/xen: Shutdown request handler
Starts a handler thread that waits for request via the
XenStore entry `control/shutdown`.
Please note that in this commit we call directly
`ukplat_terminate()`. As soon as we are ready with proper
shutdown support, we should forward the request to the
application and general libraries so that they can take
care of all needed actions for a clean shutdown (e.g.,
clean filesystem unmount).
The handler logic is derived from NEC Labs clone of Mini-OS:
https://github.com/sysml/mini-os/blob/master/shutdown.c

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>