lib/vfscore: Implement dentry cache
This commit introduces a dentry cache which keeps dentries on
standby so they are readily available if they are needed again
after closing. The changes to the dentry lookup also make it
unnecessary for dentries to be named, thus allowing vnodes that
do not have a name (e.g., epoll fd) to be created without name
so they do not waste space in the cache and the lookup hash
table.

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@kit.edu>