drivers/virto_ring: Fix IRQ corruption
When virtio block driver allocates descriptor to represent an I/O to the
virtio device, it may happen that an IRQ is raised in the middle of the
allocation of descriptor entries. As the IRQ handler in the block device
returns descriptor entries, the descriptor free list can be corrupted.
This commit uses atomic instruction to implement a lockless descriptor
list update that works with IRQs enabled.

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@kit.edu>