drivers/virtio_*: Fix unhandled IRQ
It may happen that depending on timing there is no new data in the
virtio queue, however, an IRQ is still pending. When this pending IRQ
is handled, the driver reports back that it did not handle the request
although in fact resetting the IRQ signal bit in the virtio device is
enough to acknowlegde that an interrupt for the device has been
received and handled (albeit without having to do any additional
work). This will lead to a false Unhandled IRQ error message on
the console. This commit is a first dirty hack to just always
acknowledge IRQs if the status bit is set, thereby avoiding the
error message.

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@kit.edu>