lib/ukschedcoop: Re-factor scheduler instance creation and instance object
This commit re-factors the scheduler creation call by renaming if from
`uk_schedcoop_init()` to `uk_schedcoop_create()` because the scheduler
is allocating an instance object. Additionally, the memory layout is
simplified by embedding `struct uk_sched` to `struct uksched_coop`.
The `struct uk_sched_coop` is treated privately to the scheduler
implementation. The allocator that should be utilized by the scheduler
for allocation is also stored on the generic `struct uk_sched`.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>