plat/xen: make timestamps relative to boot
In Xen, UK uses the system timestamp from the shared info
to measure time for the monotonic clock. The system timestamp represents
the uptime of the underlying Xen host. This makes it difficult to
measure the actual time needed to boot the system because we then need
to get a reference time that indicates when the VM started. This commit
changes the Xen time code to measure this reference during system setup
and apply it to all timestamps internally. This way, the monotonic
clock represents (roughly) the time since UK has been started, which is
inline with how timestamps on KVM work.

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@kit.edu>