lib/uksched: Introduce `uk_thread_uktls_var()`
This commit introduces the ability to modify TLS variables of an
foreign thread. The TLS is used as an extensible way to implement
Unikraft's TCB. Such an access method is required to allow
initialization of foreign threads. The implemented method works
only with static TLSes for now.
In addition, changing the currently active TLS temporarily is
supported as well: On thread switches, the CPU register containing
the TLS pointer is saved and restored. For this purpose, the
reference of the original TLS is stored in the `uktlsp` field
within `uk_thread`.
This mode can also be utilized for binary compatible modes where
externally built binaries are referring to a different TLS than
code that was built with Unikraft.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>