plat/xen/x86: Rewrite static memory initialization
When the system starts in paravirtual mode, Xen has set up a page table
that maps the kernel image, initrd, and the page table pages. To make
all physical memory accessible the platform code extends the page
table page-by-page. Since the page table is currently active, we cannot
just do memory writes to set PTEs but must use hypercalls. In addition,
new page table pages need to be set read-only before they can be
linked into the hierarchy, adding even more hypercalls.

This approach leads to prohibitive high costs for setting up the initial
virtual memory address space. This commit rewrites the entire memory
initialization with a focus on saving hypercalls and thus reducing boot
time considerably.

The idea is to build a new page table in memory before actually
advertising it to Xen. We mostly copy the PTEs from the startup
page table and create new ones for the unmapped physical memory.
However, when the page table is made active, Xen walks the whole
table to make sure that all PTEs respect memory isolation. Although
this is already faster than working on the active page table, it is still
not really fast. We therefore limit the amount of memory we actually
map. For all other memory, we build the page table hierarchy and
create the PTE, but leave the present bit disabled. We use a page
fault handler to lazily map these entries when they are actually
needed. Since the page table hierarchy is already set up at this point
we then only need a single hypercall to resolve the page fault.

A limitation of the new code is that it needs the page table pages to
be already mapped when the system starts. We achieve this by
reserving space in the .bss section. This of course means we have to
limit the maximum amount of supported memory at compile-time
and potentially waste space, when the system does not exhaust the
maximum. However, this limitation is bearable as paging should
be enabled for fully-fledged memory support.

Tighly integrated into the Xen memory management code is the
allocation and mapping of grant table frames. Currently, the grant
table frames are allocated and mapped using a rather dynamic
approach involving the memory allocator, page table modifications
on the active page table, etc. although the actual number of grant
table frames is fixed at compile time. We thus remove all of this and
just allocate and map the grant table frames during initial page
table setup, thereby saving hypercalls and unecessary complexity.

Duplicate types and definitions are removed to improve integration
with the paging code.

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@kit.edu>