lib/ukmmap: Try release memory from partial frees
The current implementation of ukmmap leaks the mmap'ed memory in
munmap if the caller does not unmap the whole region. This patch adds
a remaining length field to every VMA so that the whole mmap region is
freed when the amount of memory freed in a VMA reaches the VMA's
size. This is a hack to reduce the amount of leaked memory.

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@kit.edu>