arch/libcontext: aarch64: Use SP pointer instead of X2 when storing ctx
Let's consider context switch routine

ENTRY(asm_sw_ctx_switch)
    /* Save callee-saved registers to prevctx's stack */
    sub x2, sp, #__CALLEE_SAVED_SIZE
    stp x19, x20, [x2, #16 * 0]
    ...
    stp x29, x30, [x2, #16 * 5]
    (point 1) <-----------------------------------
    /*
     * Record the restore point for switch out thread to restore
     * its called-saved registers in next switch to time.
     */
    ldr x30, =restore_point
    /* Save sp and restore point to previous context */
    stp x2, x30, [x0]

When we save context, sp is not allocated (not decremented), because we use
intermediate x2 register. So at point1 the stack picture would be
  (addrs grow down)

   x19   <----- sp
   x20
    ..
   x30   <------x2

however, if the irq will occur at point 1, it will corrupt saved context
with the trapframe data, because it uses the same stack and sp was not
allocated.

Signed-off-by: Aleksandr Iashchenko <aia@opensynergy.com>