lib/syscall_shim: Use AWK rules from the build system for C sources
Uses the buiuld system's support for pre-processed source files to
simplify the library's `Makefile.uk`. The generation of these
files is only triggered when the set of system calls changes.

Because generated header files are stored in a different destination
path, the same improvement cannot performed for these files, at the
moment.

Signed-off-by: Simon Kuenzer <simon@unikraft.io>